tcp code

+++ router/sr_nat.h
@@ -80,7 +80,7 @@ int   sr_nat_destroy(struct sr_nat *nat);  /* Destroys the nat (free memory) */
 void *sr_nat_timeout(void *nat_ptr);  /* Periodic Timout */
 int is_nat_internal_iface(char *iface);
 int is_nat_external_iface(char *iface);
-int generate_unique_port(struct sr_nat *nat);
+uint16_t generate_unique_port(struct sr_nat *nat);
 /* Get the mapping associated with given external port.
    You must free the returned structure if it is not NULL. */
 struct sr_nat_mapping *sr_nat_lookup_external(struct sr_nat *nat,
